Use groups to simplify tooltips and deduplicate
Where we have duplicate unit lists a new module has been setup which declares them for us. These groups will also be used in tech cards to ensure alignment between the tooltip and the card. This also allowed for the identification of a couple of errors which have also been fixed.