Future proof card tooltips against deal size change
Previously we relied on there being no more than 4 cards in a deal, with a card in the player's inventory hover using the final slot in the array to avoid conflicting with the deal tooltips. This required changing the magic number every time the largest possible deal limit was changed. Instead the inventory tooltip now gets the first slot in the array and all the deal cards are offset by 1 so that the deal size will never conflict with it.