Fixed Workers with minions commander number
If the minion count was low enough and the Worker was the primary system opponent or a foe then it wouldn't receive extra workers. It will now always receive at least one. Due to the use of a single function this now always results in a FFA Worker getting at least 2 commanders, but this is considered acceptable.