Switch to for...of in place for for where possible
This was done previously and reverted due to eslint not supporting this in an es5 environment, and setting it up for es6 causing it to allow things PA wouldn't support. Since then a requirement for a supported ES6 feature has come up so the linter now supports ES6 elements and we might as well use them where we can.