Protect against war generation failure
This was happening when GWBreeder trimmed the AI teams list due to a lack of spawns because GWO always assumed it had boss information for every team. Vanilla handles this by checking for a boss on each team, but due to GWO sometimes only having one faction in a galaxy it doesn't make sense to allow trimming and therefore we restart war generation with a next galaxy. This implementation protects against an infinite loop of generation due to some deeper issue with galaxy generation.