Fix tech card modifiers not being applied
This was caused by each unit's associated specs not being found. This is likely some specific issue with function dependencies existing outside the scope of $.when. This is likely the same issue as seen with dealCardToSelectableAI() where $.when().then() doesn't wait for associated functions to return a value before continuing, and this would indicate that may only be the case where the function is outside of $.when()'s scope.