Guardians change personality to match player tech
Before the Guardians inherited the unit preferences of the faction they replaced, now they check the player's tech. This doesn't factor in individual unit techs, just the general techs. This is due to ease of implementation, and to suit the more general unit compositions the AI brings. A future enhancement could be to have the AI specifically favour units with individual techs.