Support enemy and ally AIs using different brains
Due to the underlying code working on the assumption that only a single brain was in use, this required a fairly sizable rewrite of referee_ai.js, along with various supporting files. I used this as an opportunity to refactor the code to be more maintainable, though it still needs work to be more readable.

Expanding beyond two brains will need further work.